Addressed Java8 javadoc issues. Altered tests for Windows platform.
