+++ src/test/java/org/jbake/app/FileUtilTest.java
@@ -2,9 +2,11 @@ package org.jbake.app;
 
 import static org.assertj.core.api.Assertions.*;
 
+import org.jbake.TestUtils;
 import org.junit.Test;
 
 import java.io.File;
+import java.io.IOException;
 import java.net.URL;
 
 /**
@@ -20,13 +22,15 @@ public class FileUtilTest {
     }
     
     @Test
-    public void testFileIgnore(){
+    public void testFileIgnore() throws IOException, InterruptedException {
     	URL sourceUrl = this.getClass().getResource("/fixture");
     	File contentFolder = new File(sourceUrl.getFile(), "content");
     	
     	//Without filter, make sure ignorable file is selected
     	File files1[] = contentFolder.listFiles();
     	assertThat(files1).contains(new File(contentFolder, ".ignorablefile.html"));
+    	
+    	TestUtils.hideAssets(contentFolder);
     	
     	//When using filter, ignorable file should not be selected
     	File files2[] = contentFolder.listFiles(FileUtil.getFileFilter());
@@ -36,7 +40,7 @@ public class FileUtilTest {
     @Test
     public void testFolderIgnore(){
     	URL sourceUrl = this.getClass().getResource("/fixture");
-    	File contentFolder = new File(sourceUrl.getFile() + "/content/blog");
+    	File contentFolder = new File(sourceUrl.getFile());
     	
     	//Without filter, make sure ignorable folder is selected
     	File files1[] = contentFolder.listFiles();
